Phase 2 of #129: reshape :core:tunnel to jvm-only layout
- Delete src/androidMain/: its AndroidManifest.xml was contributing
  INTERNET/WAKE_LOCK/FOREGROUND_SERVICE to the merge, and its Kotlin
  files were stale duplicates that were not compiled (the module already
  had no Android plugin applied).
- Move the three permissions that actually need to reach the merged APK
  into :work/src/main/AndroidManifest.xml (WAKE_LOCK joins the existing
  FOREGROUND_SERVICE* declarations there; INTERNET remains on :app).
- Update the stale comment in :app manifest describing the permission
  contributors.
- Rename src/commonMain -> src/jvmMain and merge src/commonTest/ into
  existing src/jvmTest/ via git mv to preserve blame.
- Collapse CsrGenerator.kt's expect/actual split into a single concrete
  JVM class (delete CsrGenerator.jvm.kt): expect/actual in the same
  module is rejected by the compiler, and the module is now JVM-only.
- Update build.gradle.kts sourceSets to reference jvmMain/jvmTest and
  fold the former jvmMain extras (ktor-client-cio) into the single
  jvmMain block.
