--- core/tunnel/src/androidMain/kotlin/com/rousecontext/tunnel/TunnelClient.kt
@@ -1,87 +1,0 @@
-package com.rousecontext.tunnel
-
-import java.io.InputStream
-import java.io.OutputStream
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.SharedFlow
-import kotlinx.coroutines.flow.StateFlow
-
-/**
- * Connection state machine for the mux tunnel.
- */
-enum class TunnelState {
-    /** No mux connection. Waiting for FCM or explicit connect. */
-    DISCONNECTED,
-
-    /** WebSocket handshake + mTLS in progress. */
-    CONNECTING,
-
-    /** Mux open, no active streams. */
-    CONNECTED,
-
-    /** Mux open, one or more active streams. */
-    ACTIVE,
-
-    /** Teardown in progress, sending CLOSE for active streams. */
-    DISCONNECTING,
-}
-
-/**
- * Errors emitted by the tunnel for the app layer to handle.
- */
-sealed interface TunnelError {
-    data object ConnectionFailed : TunnelError
-    data object AuthRejected : TunnelError
-    data object CertExpired : TunnelError
-    data class StreamError(val streamId: Int, val code: Int) : TunnelError
-    data class Unexpected(val message: String) : TunnelError
-}
-
-/**
- * A demultiplexed stream within the mux connection.
- *
- * Each stream represents a single MCP client session. The tunnel provides
- * plaintext I/O (post-TLS termination); the app layer wires these to McpSession.
- */
-interface MuxStream {
-    /** Mux routing ID (per-connection, assigned by relay). */
-    val streamId: Int
-
-    /** UUID generated by tunnel for audit log and notifications. */
-    val sessionId: String
-
-    /** SNI hostname from the connecting client. */
-    val sniHostname: String
-
-    /** Plaintext input (post-TLS termination by tunnel). */
-    val input: InputStream
-
-    /** Plaintext output (tunnel encrypts on the way out). */
-    val output: OutputStream
-
-    /** Close this stream, sending a CLOSE frame to the relay. */
-    suspend fun close()
-}
-
-/**
- * Tunnel client interface. Manages the mux WebSocket connection to the relay
- * and demultiplexes incoming streams.
- *
- * The tunnel knows nothing about MCP. It provides raw I/O streams.
- */
-interface TunnelClient {
-    /** Current connection + stream state. */
-    val state: StateFlow<TunnelState>
-
-    /** Errors emitted for app to handle (notifications, audit, UI). */
-    val errors: SharedFlow<TunnelError>
-
-    /** Connect to relay (called by foreground service on FCM wakeup). */
-    suspend fun connect()
-
-    /** Disconnect from relay. */
-    suspend fun disconnect()
-
-    /** Emitted when relay sends OPEN -- app should create McpSession for this stream. */
-    val incomingSessions: Flow<MuxStream>
-}
