--- core/tunnel/src/androidMain/kotlin/com/rousecontext/tunnel/TunnelService.kt
@@ -1,11 +1,0 @@
-package com.rousecontext.tunnel
-
-import androidx.lifecycle.LifecycleService
-
-/**
- * Core tunnel service handling FCM wakeup, ACME certificate provisioning,
- * SNI passthrough relay, session lifecycle, and wakelock management.
- */
-class TunnelService : LifecycleService() {
-    // TODO: implement tunnel lifecycle
-}
