+++ core/tunnel/src/jvmMain/kotlin/com/rousecontext/tunnel/TunnelError.kt
